move delivery date check into placeOrder function
